ai-decision-system/requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
dash==2.16.1
dash-bootstrap-components==1.5.0
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
joblib==1.3.2
flask==3.0.0
flask-cors==4.0.0
pymysql==1.1.0
sqlalchemy==2.0.23
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6 